+++ src/Language/AST/TypeSystemDefinitionNode.php
@@ -6,10 +6,7 @@ namespace GraphQL\Language\AST;
  * export type TypeSystemDefinitionNode =
  * | SchemaDefinitionNode
  * | TypeDefinitionNode
- * | TypeExtensionNode
  * | DirectiveDefinitionNode.
- *
- * @property NameNode $name
  */
 interface TypeSystemDefinitionNode extends DefinitionNode
 {